better optimization.Use the common method of zorder
