--- spark/v3.3/spark/src/main/java/org/apache/iceberg/spark/actions/HilbertCurveUtils.java
@@ -1,50 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.iceberg.spark.actions;
-
-import java.math.BigInteger;
-import org.davidmoten.hilbert.HilbertCurve;
-
-/** Utils for Hilbert Curve. */
-public class HilbertCurveUtils {
-  public static byte[] indexBytes(HilbertCurve hilbertCurve, long[] points, int paddingNum) {
-    BigInteger index = hilbertCurve.index(points);
-    return paddingToNByte(index.toByteArray(), paddingNum);
-  }
-
-  private HilbertCurveUtils() {}
-
-  public static byte[] paddingToNByte(byte[] data, int paddingNum) {
-    if (data.length == paddingNum) {
-      return data;
-    }
-    if (data.length > paddingNum) {
-      byte[] result = new byte[paddingNum];
-      System.arraycopy(data, 0, result, 0, paddingNum);
-      return result;
-    }
-    int paddingSize = paddingNum - data.length;
-    byte[] result = new byte[paddingNum];
-    for (int i = 0; i < paddingSize; i++) {
-      result[i] = 0;
-    }
-    System.arraycopy(data, 0, result, paddingSize, data.length);
-    return result;
-  }
-}
